requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.6
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
